Create ESP32-deauthing and evil twin
The turkish version of an already-made project: The wifi which esp32 made's password will be yours. The networks esp32 saw will be on the Wifi enter portal. You'll select network you want to do something with it, then select what to do. Your choices are evil twin and deauthing. let me be honest, deauthing doesnt work. So there is a choice left, evil twin. 

If you dont know what is evil twin, this paragraph is for you. Evil twin makes a new network whichs name is same as the old and removes users from the previous network. If user is a knowless person, it will choose the network we created. On the join portal, network gonna say "We saw a porblem blablabla so we need your wifi password again" the stupid user enters it, and what is it? The password writes on your esp32. You own that network.